fix(constitution): @workkit/vitest-config helper as plain ESM (Node CI compat)
CI Node runtime cannot load .ts files directly; the helper module needed
to be plain JavaScript (with .d.ts sidecar for TS consumers). Switched
tooling/vitest/index.ts → tooling/vitest/index.js + tooling/vitest/index.d.ts;
package.json now points at the .js with explicit exports map.

Local bun-driven runs masked this by handling .ts natively. CI failure on
#41 (mail and others) traced to ERR_UNKNOWN_FILE_EXTENSION on the helper.

No API change. Drive-by biome auto-format on packages/approval/vitest.config.ts.

Refs #30.
