+++ packages/approval/vitest.config.ts
@@ -1,5 +1,5 @@
 import { defineWorkkitVitest } from "@workkit/vitest-config";
 
 export default defineWorkkitVitest({
-	include: ["tests/**/*.test.ts"]
+	include: ["tests/**/*.test.ts"],
 });
