+++ tooling/vitest/package.json
@@ -3,6 +3,12 @@
 	"version": "0.0.0",
 	"private": true,
 	"type": "module",
-	"main": "./index.ts",
-	"types": "./index.ts"
+	"main": "./index.js",
+	"types": "./index.d.ts",
+	"exports": {
+		".": {
+			"types": "./index.d.ts",
+			"default": "./index.js"
+		}
+	}
 }
